Authorize Users
determine the role based on the accepted token

HomeController of PI02 -
1) added claims type without  roles added (admin)
result when consumeToken2 ( from PC01 ) is called, it is not authorized
to use the controller.

2) when claimstype.role is added (admin)
it is now authorized to use the controller.

Conclusion: we can now determine and use the claims role depending on
the token they have.